Little modification
I have a problem ingame... The pluggin won't work for me, I can't see my damage (I'm in french). After connecting, I have a message warning me about an error in line 400 of this file. I found this : it was a [code] at the begenning of the line... After deleted it, it work fine.
I'm new with code but if this can help...
Thanks for your work.